==> 01_TextPreprocessing_NaverMovieReview_Lecture2.ipynb <==
# 네이버 영화 리뷰 데이터에 대한 이해와 전처리

1. 네이버 영화 리뷰 데이터셋 로딩 (머신러닝을 위하여 train_set과 test_set으로 분리되어 있음)
1. 데이터 정제 : 결측치 제거, 중복값 제거, 한글과 공백을 제외한 문자 제거
1. 토큰화 및 불용어 제거
1. 전처리된 데이터 저장
1. 데이터 시각화 : 빈도수, WordCloud

## 1. 리뷰 데이터 로딩, 데이터 확인
* pandas.read_table

# 훈련용 리뷰 데이터, 테스트용 리뷰 데이터 로딩
import pandas as pd
train_data = pd.read_table("DataSet_NaverMovieReview/ratings_train.txt")
test_data = pd.read_table("DataSet_NaverMovieReview/ratings_test.txt")

# 훈련용 리뷰 개수 출력
len(train_data)

# 훈련용 데이터 50부터 20개 출력
train_data[50:70]

# 테스트용 리뷰 개수 출력
len(test_data)

# 테스트용 리뷰 데이터 앞부분 확인
test_data.head()

## 2. 데이터 정제(Cleaning)

### 결측치 제거 (test_data, train_data)
* dataframe.isnull().any(), dataframe.isnull().values.any(): 하나라도 True가 있는지 확인
* dataframe.isnull().sum() : 결측치 수
* dataframe.info()
* dataframe.loc[dataframe.document.isnull()]
* dataframe.dropna(subset,keep)

#### test_data 처리

# null값이 하나라도 있는지 확인
test_data.isnull().any()

# 결측치 수 구하기
test_data.isnull().sum()

# 결측치 정보 확인
test_data.info()

# Null 값을 가진 샘플이 어느 인덱스의 위치에 존재하는지 출력
test_data.loc[test_data.document.isnull()]

# 결측치 제거
test_data = test_data.dropna()

#결과 확인
test_data.info()

#### train_data 처리

# null값이 하나라도 있는지 확인

# 결측치 수 구하기

# 결측치 정보 확인
train_data.info()

# null 값을 가진 샘플이 어느 인덱스의 위치에 존재하는지 한 번 출력


# 결측치 제거 및 결과 확인
train_data.dropna(inplace=True)

# train_data와 test_data 수 확인
len(test_data), len(train_data)

### 중복값 제거 (test_data, train_data)
* dataframe.duplicated().any()
* dataframe.duplicated().sum()
* dataframe.nunique()
* dataframe.drop_duplicates(subset, keep)

#### test_data 처리

# document에 중복값 존재 여부 확인
test_data['document'].duplicated().any()

# document에 중복값이 몇개 있는지 확인
test_data['document'].duplicated().sum()

# 중복값이 있는 데이터 확인
test_data.loc[test_data['document'].duplicated()]

# document 열과 label 열의 중복을 제외한 값의 개수
test_data['document'].nunique()

# document 열의 중복 제거
test_data.drop_duplicates(subset=['document'], inplace=True)

# 결과 확인
test_data['document'].duplicated().sum()

len(test_data)

#### train_data 처리

# document에 중복값이 있는지 여부 확인


# document에 중복값이 몇개 있는지 확인

# 중복값이 있는 데이터 확인


# document 열과 label 열의 중복을 제외한 값의 개수


# document 열의 중복 제거
train_data.drop_duplicates(subset=['document'], inplace=True)

# 결과 확인


# 전체 데이터 수 확인
len(train_data), len(test_data)

### 한글과 공백을 제외하고 모두 제거
* document에서 한글과 공백을 제외하고 제거 : [^ ㄱ-ㅣ가-힣]+
* document에 공백만 남은 경우 None를 변경하고 결측치 제거 : '^ +'

#### test_data 처리

# document의 첫 5개 데이터 보기
test_data.head()

# 한글과 공백을 제외하고 제거
import re
test_data['document'] = test_data['document'].apply(lambda x: re.sub(r'[^ ㄱ-ㅣ가-힣]+', "", x))
# document의 첫 5개 데이터 보기
test_data.head()

test_data.info()

# document에 공백만 남은 경우 None를 변경하고 결측치 제거
# white space 데이터를 empty value로 변경
test_data['document'] = test_data['document'].apply(lambda x:re.sub('^ +', "", x))
# empty value를 None으로 변경
test_data['document'].replace("", None, inplace=True)

# 결측치 수 확인
test_data.info()
test_data['document'].isnull().sum()

# 결측치 제거
test_data.dropna(inplace=True)

# document의 첫 5개 결과 보기
test_data.info()

#### train_data 처리

# 한글과 공백을 제외하고 제거
import re
train_data['document'] = train_data['document'].apply(lambda x: re.sub(r'[^ ㄱ-ㅣ가-힣]+', "", x))
# document의 첫 5개 데이터 보기
train_data.head()

# document에 공백만 남은 경우 None를 변경하고 결측치 제거
## white space 데이터를 empty value로 변경
train_data['document'] = train_data['document'].apply(lambda x:re.sub('^ +', "", x))
# empty value를 None으로 변경
train_data['document'].replace("", None, inplace=True)

# 결측치 수 확인

# 결측치 제거, 결측치 수 확인
train_data.dropna(inplace=True)
train_data.info()

# 전체 데이터수 확인
len(train_data), len(test_data)

temp_train_data = train_data.copy()
temp_train_data.drop(columns=['tokenized'], inplace=True)

temp_train_data.head()

temp_test_data = test_data.copy()
temp_test_data.drop(columns=['tokenized'], inplace=True)
temp_test_data.head()

temp_train_data.to_csv('DataSet_NaverMovieReview/train_data_prep_sent.csv', index=False)
temp_test_data.to_csv('DataSet_NaverMovieReview/test_data_prep_sent.csv', index=False)

train_data.to_csv('DataSet_NaverMovieReview/train_data_prep_sent.csv', index=False)
test_data.to_csv('DataSet_NaverMovieReview/test_data_prep_sent.csv', index=False)

## 3. 토큰화, 불용어 제거
* konlpy의 Okt로 토큰화
* 불용어 제거

#### test_data 처리

# 불용어 정의
stopwords = ['의','가','이','은','들','는','좀','잘','걍','과','도','를','으로','자','에','와','한','하다']

testest = test_data[:5].copy()

# 토큰화한 후 불용어를 제거하여 데이터 리스트 생성
from konlpy.tag import Okt
okt = Okt()
new_test_data = []
for sentence in testest['document']:
    tok_sent = okt.morphs(sentence, stem=True) # 토큰화
    sw_rm_sent = [word for word in tok_sent if word not in stopwords]
    new_test_data.append(sw_rm_sent)
new_test_data

# dataframe에 'tokenized' 컬럼 추가하여 불용어를 제거한 토큰화 결과 저장
from konlpy.tag import Okt
okt = Okt()
stopwords = ['의','가','이','은','들','는','좀','잘','걍','과','도','를','으로','자','에','와','한','하다']

def okt_sw_rm_tokenize(sentence):
    tok_sent = okt.morphs(sentence, stem=True) # 토큰화
    sw_rm_sent = [word for word in tok_sent if word not in stopwords]
    return sw_rm_sent

test_data['tokenized'] = test_data['document'].apply(okt_sw_rm_tokenize)

# data 확인
test_data.head()

#### train_data 처리

# dataframe에 'tokenized' 컬럼 추가하여 불용어를 제거한 토큰화 결과 저장
train_data['tokenized'] = train_data['document'].apply(okt_sw_rm_tokenize)

train_data.head()

## 4. 전처리된 데이터 저장과 로딩
* dataframe.to_csv(filename, index=False)
* dataframe = pandas.read_csv(filename)

# 저장
test_data.reset_index(inplace=True)
train_data.reset_index(inplace=True)
test_data.to_json('DataSet_NaverMovieReview/test_data_prep.json')
train_data.to_json('DataSet_NaverMovieReview/train_data_prep.json')

# 로딩
test_data_loaded = pd.read_json('DataSet_NaverMovieReview/test_data_prep.json')
train_data_loaded = pd.read_json('DataSet_NaverMovieReview/train_data_prep.json')

test_data.reset_index(inplace=True)

# 원본과 저장한 후 로딩한 자료 비교해보기
test_data['document'].equals(test_data_loaded['document'])

test_data.head()

test_data_loaded.info()

## 5. 빈도수 및 WordCloud 확인 -> 불용어 등 추가
* 빈도수 구하기 : Counter
    - from collections import Counter
    - counter.most_common(number)

ttt = test_data[:5].copy()

target_word = sum(list(test_data['tokenized']), [])
target_word

target_word = sum(list(ttt['tokenized']), [])
target_word
