==> 04_SA_MachineLearning_NaverMovie.ipynb <==
# 머신러닝 기반 네이버 영화 리뷰 감성 분석

# 데이터셋 로딩하여 사용


## Decision Tree
- criterion : 분할 기준 ['gini', 'entropy']
- splitter : 노드를 분할하는 데 사용되는 전략 ['best', 'random']
- max_depth': 트리의 최대 깊이
- min_samples_split : 노드 분할을 위한 최소 샘플 수 (2 ~)
- min_samples_leaf : 리프 노드에 필요한 최소 샘플 수 (1 ~)
- max_features : 각 분할에서 고려할 특성의 최대 수
    - [None, 'auto', 'sqrt', 'log2']
- class_weight: [None, 'balanced']
- random_state : 무작위성을 조절하기 위한 시드 값
    - 같은 시드로 모델을 학습하면 항상 동일한 결과를 얻을 수 있음    

## 랜덤 포레스트 (Random Forrest)
- n_estimators : Dectsion Tree의 수 (기본 100개)
- bootstrap (True)
  - Bootstrap Sampling : 학습 데이터에서 데이터를 중복해서 샘플링하는 방식
- n_jobs (None) : 병렬처리에 사용할 CPU 코어의 수

## LogisticRegression (로지스틱 회귀분석)
- penalty (default='l2'):규제 (Regularization)의 종류 지정
    - 'l1', 'l2', 'elasticnet', 'none'
    - 모델의 복잡도를 제어하고, 과적합(Overfitting)을 방지하는 기술
    - 모델의 가중치(weight)나 계수(coefficient)에 페널티를 부과하여 모델이 너무 복잡해지지 않도록 제한하는 역할
- C (default=1.0): 규제 강도 조절. 값이 작을수록 강한 규제가 적용
- fit_intercept (default=True):
- max_iter (default=100): 최적화 알고리즘이 수렴할 때까지 반복하는 최대 횟수
- random_state (default=None): 무작위성을 조절하기 위한 시드 값
- tol (default=1e-4): 수렴 기준으로, 손실 함수의 변화가 이 값보다 작을 때 최적화를 중단
- verbose (default=0): 최적화 과정 중에 출력을 활성화하려면 1 이상의 값을 지정
- warm_start (default=False): 이전에 학습된 모델을 초기화하지 않고 이어서 학습할 때 True로 설정

## SVM 모델
* penalty (default='l2')
* C (default=1.0):
* loss (default='squared_hinge'): 손실 함수 지정
    - 'hinge' : 힌지 손실, 'squared_hinge' : 힌지 손실의 제곱
* random_state
* max_iter
* tol

## 퍼셉트론 모델 (인공신경망)
* penalty (default=None):
* alpha (default=0.0001): 규제 강도를 조절. 높을 수록 강한 규제
* max_iter (default=1000)
* tol (default=1e-3)
* shuffle (default=True): 각 에포크(epoch)마다 훈련 데이터를 섞을지 여부
* eta0 (default=1.0): 학습률
* random_state (default=0)

## 나이브 베이지안
- 데이터가 가우시안 분포를 따른다고 가정을 따르면 일반적으로 데이터의 특성에 맞춰 잘 동작

## K-Nearest Neighbor (KNN)
- n_neighbors (default=5): 이웃의 개수(k)
- weights (default='uniform'):예측에 사용되는 가중치를 지정
    - uniform : 균일한 가중치
    - distance :
- algorithm (default='auto'): 이웃을 찾기 위해 사용되는 알고리즘
    - 'auto', 'ball_tree', 'kd_tree', 'brute'
- leaf_size (default=30): BallTree 또는 KDTree 알고리즘을 사용할 경우 잎 노드의 크기
- p (default=2): 거리 측정에 사용되는 매개변수
    - p=1 : 맨해튼 거리, p=2 : 유클리디안 거리
- metric (default='minkowski'):거리 측정에 사용되는 메트릭
    - 'minkowski'를 선택하면 p 매개변수에 따라 선택
- metric_params (default=None): 사용자 지정 거리 메트릭 함수에 대한 추가 매개변수
- n_jobs (default=None): 이웃을 검색하는 동안 사용되는 CPU 수. None = 1개
