==> 05_SA_LSTM_IntegerEncoding_NaverMovie.ipynb <==
# 네이버 영화 리뷰 감성 분석 (LSTM)
* Integer Encoding 적용한 데이터셋 로딩하여 사용

# 특징벡터화된 데이터 셋 로딩

# Integer Coding을 사용하는 vectorizer 로딩


# 1. LSTM

* EarlyStopping (callback) : 모델의 훈련을 조기에 중지. 특정 지표가 더 이상 개선되지 않을 때 훈련을 중지
    - monitor: 모니터할 지표
    - mode: 개선 여부를 결정할 기준. 'min' = 지표가 감소할 때 중지, 'max' = 지표가 증가할 때 중지
    - verbose: 중지되었을 때 메시지를 출력할지 여부
    - patience: 지정된 횟수 동안 지표가 개선되지 않을 경우 훈련이 중지
* ModelCheckpoint(callback) : 훈련 중에 모델의 가중치를 저장. 주로 검증 데이터에서 지정된 지표가 개선될 때만 모델을 저장하는 데 사용
    - filepath: 모델을 저장할 파일 경로
    - monitor: 모니터할 지표
    - mode: 개선 여부를 결정할 기준. 'min' = 지표가 감소할 때 중지, 'max' = 지표가 증가할 때 중지
    - verbose: 중지되었을 때 메시지를 출력할지 여부
    - patience: 지정된 횟수 동안 지표가 개선되지 않을 경우 훈련이 중지
    - save_best_only: 개선된 모델만 저장할지 여부

### keras 한국어 tutorial https://www.tensorflow.org/tutorials?hl=ko

### 1-1. 필요한 Callback 지정

# 훈련 조기 중지, 훈련 중 모델 저장에 대한 callback 지정


### 1-2. 신경망 구성
- 모델 유형 결정 : Sequential, Pre-trained Models 등
- 워드 임베딩 레이어 추가
- 원하는 유형 레이어 추가
    - Embedding Layer
    - Dense Layer (Fully Connected Layer)
    - Recurrent Layer : LSTM, GRU
    - Activation Layer
    - Convolutional Layer
    - Pooling Layer : MaxPooling, AveragePooling
    - Flatten Layer
    - Dropout Layer
    - Batch Normalization Layer

from tensorflow.keras.layers import Embedding, Dense, LSTM
from tensorflow.keras.models import Sequential

# 순차적 모델을 생성하기 위한 객체 초기화

# 임베딩 레이어 추가, vocab_size와 임베딩 차원 설정

# LSTM 레이어 추가, 은닉층의 갯수 설정

# 이진 분류를 위한 fully connected layer, 출력 크기 1, 활성화 함수 : sigmoid


### 1-3. 모델 학습에 필요한 최적화 알고리즘, 손실함수, 평가 지표를 저장
- 최적화 알고리즘 : RMSprop (Adagrad의 변형. 학습률을 동적으로 조절)
- 손실함수 : binary_crossentropy (이진 분류 문제에 주로 사용)
- 평가지표 : 정확도

### 1-4. 훈련 진행

### 1-5. 훈련 중 저장된 모델을 로딩하여 평가

# 2. 문장을 입력 받아 감성 분석 해보기

analyze_sentiment('이 영화 개꿀잼 ㅋㅋㅋ')

analyze_sentiment('이 영화 핵노잼 ㅠㅠ')

analyze_sentiment('이딴게 영화냐 ㅉㅉ')

analyze_sentiment('감독 뭐하는 놈이냐?')

analyze_sentiment('와 개쩐다 정말 세계관 최강자들의 영화다')
